+++ CHANGES.md
@@ -93,6 +93,11 @@ CHANGES
     * Added `CertificateThumbprint` field.
     * Added `DpopKeyThumbprint` field.
 
+- `UserInfoRequest` struct
+    * Added `Dpop` field.
+    * Added `Htm` field.
+    * Added `Htu` field.
+
 - New components
     * `PushedAuthReqAction` type and some of its constants
     * `PushedAuthReqRequest` struct